backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-rtmidi==1.5.8
python-osc==1.8.3

